% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_versions.R
\name{compare_sample_sizes}
\alias{compare_sample_sizes}
\title{Compare sample sizes by stratum between two versions}
\usage{
compare_sample_sizes(
  dataset_a,
  dataset_b,
  column_a,
  column_b,
  join_by = "stratum"
)
}
\arguments{
\item{dataset_a}{Dataset containing summarized data of source a.}

\item{dataset_b}{Dataset containing summarized data of source b.}

\item{column_a}{Column in dataset_a containing the summarized value.}

\item{column_b}{Column in dataset_b containing the summarized value.}

\item{join_by}{Column used to join both datasets. Default "stratum".}
}
\value{
Tibble with columns stratum and n (= number of unique samples)
}
\description{
Returns a merged dataframe of the two input tables with an extra column
\code{delta} containing the difference between the input columns (b-a).
}
