% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_2.R
\name{create_graph_comparison}
\alias{create_graph_comparison}
\title{Create a comparison plot of sample sizes between two versions}
\usage{
create_graph_comparison(version_a, version_b, scheme, hydr_class, top_n = NULL)
}
\arguments{
\item{version_a}{poc folder name (e.g. "poc_0.13.1").}

\item{version_b}{poc folder name (e.g. "poc_0.14.0").}

\item{scheme}{Meetnet identifier (e.g. "GW_03.3")}

\item{hydr_class}{Waterhuishoudingstypes identifier (e.g. c("HC1","HC12"))}

\item{top_n}{Optional integer. If provided, keeps only the top \code{top_n} groups.}
}
\value{
A ggplot object.
}
\description{
Helper function to perform assignment 2 in one run.
}
