% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_all_poc_spatial_samples}
\alias{download_all_poc_spatial_samples}
\title{Download all spatial_samples.csv files from the PAS Drive folder}
\usage{
download_all_poc_spatial_samples(
  drive_folder_id = "1gzrB-5AG-KYHmiQUThyTEhpsboMPHXeT",
  samples_folder_name = "samples",
  samples_file_name = "spatial_samples.csv",
  dest_dir = "inst/extdata"
)
}
\arguments{
\item{drive_folder_id}{Google drive folder where the poc data is stored.}

\item{samples_folder_name}{The folder in which the samples file is saved. Default "samples".}

\item{samples_file_name}{The name of the spatial samples file. Default "spatial_samples.csv".}

\item{dest_dir}{Local directory where the downloaded data will be saved. Default "inst/extdata".}
}
\description{
Downloads every \verb{poc_*} folder that contains a \code{samples/spatial_samples.csv}
file and saves it into the local \verb{data/} directory.
}
