% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_spatial_samples}
\alias{filter_spatial_samples}
\title{Filter spatial samples by scheme and hydr_class.}
\usage{
filter_spatial_samples(spatial_samples, scheme, hydr_class)
}
\arguments{
\item{spatial_samples}{A data.frame containing spatial samples data.}

\item{scheme}{The scheme ('meetnet') you want to retrieve data for.}

\item{hydr_class}{A list of hydr_classes ('waterhuishoudingstypes') you want to retrieve data for.}
}
\value{
The filtered input table for the given scheme and hydr_class.
}
\description{
The hydr_class per type is obtained via \code{n2khab::read_types()[, c("type","hydr_class")]}.
More info can be found on: https://github.com/inbo/n2khab
}
\details{
The hydr_class data is matched to the input dataframe by \code{type}, which can be
retrieved from \code{stratum} by extracting the part before the first underscore
(if present). If not present, type is equal to stratum.
}
