% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_versions.R
\name{plot_sample_size_dumbbell}
\alias{plot_sample_size_dumbbell}
\title{Dumbbell plot for comparing two versions by stratum.}
\usage{
plot_sample_size_dumbbell(
  dataset,
  column_a,
  column_b,
  column_delta = "delta",
  comparison_by = "stratum",
  label_a = NULL,
  label_b = NULL,
  top_n = NULL
)
}
\arguments{
\item{dataset}{A dataframe with all necessary info.}

\item{column_a}{Column containing sample size for version a.}

\item{column_b}{Column containing sample size for version b.}

\item{column_delta}{Column containing difference between column a and b. Default "delta".}

\item{comparison_by}{Column used to group the plot. Default "stratum".}

\item{label_a}{Label for version a. Defaults to column_a name.}

\item{label_b}{Label for version b. Defaults to column_b name.}

\item{top_n}{Optional integer: keep only top_n for readability (by abs(delta)).}
}
\value{
A ggplot object.
}
\description{
Dumbbell plot for comparing two versions by stratum.
}
