% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_1.R
\name{read_and_filter_dataset}
\alias{read_and_filter_dataset}
\title{Read and Filter spatial samples by scheme and hydr_class.}
\usage{
read_and_filter_dataset(version, scheme, hydr_class)
}
\arguments{
\item{version}{poc folder name (e.g. "poc_0.14.0").}

\item{scheme}{Meetnet identifier (e.g. "GW_03.3")}

\item{hydr_class}{Waterhuishoudingstypes identifier (e.g. c("HC1","HC12"))}
}
\value{
Filtered tibble of spatial samples.
}
\description{
Helper function to perform assignment 1 in one run.
}
