% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_spatial_samples}
\alias{read_spatial_samples}
\title{Read spatial_samples.csv for a given poc folder.}
\usage{
read_spatial_samples(
  poc_folder_name,
  base_dir = "inst/extdata",
  samples_folder_name = "samples",
  samples_file_name = "spatial_samples.csv",
  required_cols = c("scheme", "stratum", "grts_address")
)
}
\arguments{
\item{poc_folder_name}{Folder name of the poc version, (e.g. "poc_0.14.0").}

\item{base_dir}{Base folder containing the poc data. Default "inst/extdata".}

\item{samples_folder_name}{The folder in which the samples file is saved. Default "samples".}

\item{samples_file_name}{The name of the spatial samples file. Default "spatial_samples.csv".}

\item{required_cols}{List of required columns in the input file.}
}
\value{
A table containing the spatial samples data.
}
\description{
Expected path: <data_dir>/<poc_folder_name>/samples/spatial_samples.csv
(If needed the samples folder name and file name can be adjusted.)
}
