% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_version.R
\name{summarise_sample_sizes}
\alias{summarise_sample_sizes}
\title{Summarise sample sizes by stratum}
\usage{
summarise_sample_sizes(
  spatial_samples,
  group_by = "stratum",
  output_column = "n"
)
}
\arguments{
\item{spatial_samples}{Dataset containing spatial samples data with required columns 'stratum' and 'grts_address'.}

\item{group_by}{Column name used to summarize the data. Default "stratum".}

\item{output_column}{Column name for the output column (sample size). Default "n".}
}
\value{
Tibble with columns stratum and n (= number of unique samples)
}
\description{
Sample size = number of unique grts_address values by stratum.
}
